+++ cmd/gardener/gardener_integration_test.go
@@ -3,23 +3,11 @@
 package main
 
 import (
-	"os"
 	"testing"
 
 	"cloud.google.com/go/datastore"
 	"golang.org/x/net/context"
-	"google.golang.org/api/option"
 )
-
-func Options() []option.ClientOption {
-	opts := []option.ClientOption{}
-	if os.Getenv("TRAVIS") != "" {
-		authOpt := option.WithCredentialsFile("../../travis-testing.key")
-		opts = append(opts, authOpt)
-	}
-
-	return opts
-}
 
 func TestStartup(t *testing.T) {
 	client, err := datastore.NewClient(context.Background(), "mlab-testing", Options()...)
