+++ cmd/gardener/gardener_test.go
@@ -3,12 +3,23 @@ package main
 import (
 	"os"
 	"testing"
+
+	"google.golang.org/api/option"
 )
 
-// Currently failing because of auth.
-func xTest_getDSClient(t *testing.T) {
+func Options() []option.ClientOption {
+	opts := []option.ClientOption{}
+	if os.Getenv("TRAVIS") != "" {
+		authOpt := option.WithCredentialsFile("../../travis-testing.key")
+		opts = append(opts, authOpt)
+	}
+
+	return opts
+}
+
+func Test_getDSClient(t *testing.T) {
 	os.Setenv("PROJECT", "mlab-testing")
-	c, err := getDSClient()
+	c, err := getDSClient(Options()...)
 	if err != nil {
 		t.Fatal(err)
 	}
